Split TestInfoProvider into files and move to separate test package
