+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/process/test/ScenarioTestDataSerDe.scala
@@ -1,7 +1,7 @@
 package pl.touk.nussknacker.ui.process.test
 
 import io.circe.parser
-import pl.touk.nussknacker.engine.definition.{PreliminaryScenarioTestData, PreliminaryScenarioTestRecord}
+import pl.touk.nussknacker.engine.definition.test.{PreliminaryScenarioTestData, PreliminaryScenarioTestRecord}
 import pl.touk.nussknacker.ui.api.TestDataSettings
 
 class ScenarioTestDataSerDe(testDataSettings: TestDataSettings) {
