+++ interpreter/src/main/scala/pl/touk/nussknacker/engine/definition/test/TestingCapabilities.scala
@@ -1,0 +1,9 @@
+package pl.touk.nussknacker.engine.definition.test
+
+import io.circe.generic.JsonCodec
+
+@JsonCodec case class TestingCapabilities(canBeTested: Boolean, canGenerateTestData: Boolean)
+
+object TestingCapabilities {
+  val Disabled: TestingCapabilities = TestingCapabilities(canBeTested = false, canGenerateTestData = false)
+}
